+++ packages/elvis-colors/CHANGELOG.md
@@ -1,11 +1,16 @@
 # Elvia Colors Changelog
 
-## 2.8.0 (??.06.23)
+## 3.0.0 (??.06.23)
+
+### Breaking changes
+
+- Static category -> Assorted category (with no common prefix)
 
 ### New feature
 
 - Building CSS variables and classes from tokens to elvis.
-- Added new theme color label: `static-brand-accent`
+- Added new theme color label: `brand-accent`, `background-element-6`, `text-4`, `text-5`
+- Dark `grey-70` has gone from `#242424` to `#262626` to match `grey` in light theme.
 
 ## 2.7.1 (07.06.23)